pyproject.toml
[project]
name = "value-proposition-canvas"
version = "1.0.0"
description = "AI-powered coaching for creating high-quality Value Proposition Canvases"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-docx>=1.1.0",
    "openai>=1.10.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "scikit-learn>=1.4.0",
    "slowapi>=0.1.9",
    # Database
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    # Authentication
    "bcrypt>=4.0.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "ui", "admin"]
